update image resize

+++ src/controllers/contributeController.ts
@@ -103,15 +103,15 @@ export default class CommentController {
         } as IContribute);
 
         var ArrayImage = [];
-      dataImages.images.forEach(item => {
-          const image = item.image;
-          const thumbnailSizes = [150, 250, 480, 720, 1200, 1280, 1920];
-          thumbnailSizes.forEach(size => {
-              console.log(image.replace('images/', `thumbnail/${size}/`).replace(/\.(png|jpg|jpeg)$/i, '.webp'));
-              ArrayImage.push(image.replace('images/', `thumbnail/${size}/`).replace(/\.(png|jpg|jpeg)$/i, '.webp'));
-          });
-          ArrayImage.push(image);
-      });
+        dataImages.images.forEach(item => {
+            const image = item.image;
+            const thumbnailSizes = [150, 250, 480, 720, 1200, 1280, 1920];
+            thumbnailSizes.forEach(size => {
+                console.log(image.replace('images/', `thumbnail/${size}/`).replace(/\.(png|jpg|jpeg)$/i, '.webp'));
+                ArrayImage.push(image.replace('images/', `thumbnail/${size}/`).replace(/\.(png|jpg|jpeg)$/i, '.webp'));
+            });
+            ArrayImage.push(image);
+        });
 
         await this.RemoveFileAny(ArrayImage);
         const data = await serviceInstance.removeContribute({
@@ -127,15 +127,19 @@ export default class CommentController {
       }
     }
 
-    public async RemoveFileAny(pathOject : any){
-      if(pathOject != null){
-        pathOject.forEach( pathLink =>{
-          fs.unlink(path.join(pathLink), (err) => {
-            if (err) {
-              console.log(err);
-            }
+    public async RemoveFileAny(pathOject: any) {
+      if (pathOject != null) {
+          pathOject.forEach(pathLink => {
+              if (fs.existsSync(path.join(pathLink))) {
+                  fs.unlink(path.join(pathLink), (err) => {
+                      if (err) {
+                          console.log(err);
+                      }
+                  });
+              } else {
+                  console.log(`File ${pathLink} does not exist.`);
+              }
           });
-        })
       }
     }
